Tipo do field

+++ models/alerta_model.py
@@ -6,5 +6,5 @@ class Alerta(gj.Document):
     nome_titulo = StringField(required=True)
     tipo_notificacao = StringField(required=True) # "SMS" ou "EMAIL"
     situacao = IntField(required=True) # -1 menor, 0 igual ou  1 maior
-    valor = DoubleField(required=True)
+    valor = DecimalField(required=True)
